Extract useCreateTask mutation hook
